add test for helper func.
